Changed the url
Changed the URL in this file to an http url rather than https.

This is because users would get a warning when they attempted to load the website with https, as the website did not support that. Changing it to http will solve the problem with these warnings, which could confuse some of the users.